# openapi.yaml
openapi: 3.0.3
info:
  title: AuthOS API
  description: |
    Comprehensive Laravel 12 authentication service API - An Auth0 alternative providing OAuth 2.0, OpenID Connect, multi-factor authentication, and single sign-on capabilities.
    
    ## Features
    - OAuth 2.0 & OpenID Connect server implementation
    - Multi-tenant organization management
    - Role-based access control (RBAC)
    - Multi-factor authentication (MFA) with TOTP
    - Real-time authentication monitoring
    - Comprehensive user and application management
    - Rate limiting with role-based adjustments
    
    ## Authentication
    This API uses OAuth 2.0 Bearer tokens for authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer {access_token}
    ```
    
    ## Rate Limiting
    Different endpoints have varying rate limits based on their usage patterns:
    - Authentication: 10 requests/minute (per IP)
    - Registration: 5 requests/hour (per IP) 
    - Standard API: 1000 requests/hour (per user)
    - Admin API: 200 requests/hour (per user)
    - Bulk Operations: 100 requests/hour (per user)
    - MFA: 20 requests/hour (per user)
    - OAuth: 20 requests/minute (per IP)
    
    Rate limits are increased for users with higher roles (Super Admin gets 5x, Organization Admin gets 3x, Application Admin gets 2x).
    
    ## Error Handling
    All API errors follow a consistent JSON format with appropriate HTTP status codes.
    
  version: 1.0.0
  contact:
...
servers:
...
paths:
...
  /auth/revoke:
    post:
      tags:
        - Authentication
      summary: Revoke access token
      description: Revoke the current access token
      operationId: revokeToken
      security:
        - bearerAuth: []
      responses:
        200:
          description: Token revoked successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Token revoked successfully"

  # OAuth 2.0 Endpoints
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        OAuth 2.0 Bearer token authentication. Include the access token in the Authorization header:
        ```
        Authorization: Bearer {access_token}
        ```
        
        ### Obtaining Access Tokens
        Access tokens can be obtained through:
        - **User Registration/Login**: `/auth/register` or `/auth/login`
        - **OAuth Authorization Code Flow**: `/oauth/authorize` → `/oauth/token`
        - **Client Credentials Flow**: `/oauth/token` (for server-to-server)
        
        ### Token Scopes
        Tokens may include different scopes that determine access levels:
        - `openid`: Basic OpenID Connect access
        - `profile`: Access to user profile information
        - `email`: Access to user email address
        - `read`: Read access to resources
        - `write`: Write access to resources
        - `admin`: Administrative access (for management endpoints)
        
        ### Token Lifetime
        Access tokens typically expire after 1 hour. Use refresh tokens to obtain new access tokens without re-authentication.
security:
  - bearerAuth: []
...